Remove antd dependencies as suggested in 4982#discussion_r162675122
